// src/clients/weatherClient.ts
export class ForecastTimeStep implements IForecastTimeStep {
    /** Forecast for a specific time */
    data!: Data;
    /** The time these forecast values are valid for. Timestamp in format YYYY-MM-DDThh:mm:ssZ (ISO 8601) */
    time!: string;

    constructor(data?: IForecastTimeStep) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.data = new Data();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.data = _data["data"] ? Data.fromJS(_data["data"]) : new Data();
            this.time = _data["time"];
        }
    }

    static fromJS(data: any): ForecastTimeStep {
        data = typeof data === 'object' ? data : {};
        let result = new ForecastTimeStep();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["data"] = this.data ? this.data.toJSON() : <any>undefined;
        data["time"] = this.time;
        return data;
    }
}

export interface IForecastTimeStep {
    /** Forecast for a specific time */
    data: Data;
    /** The time these forecast values are valid for. Timestamp in format YYYY-MM-DDThh:mm:ssZ (ISO 8601) */
    time: string;
}
// ...
export class Data implements IData {
    /** Parameters which applies to this exact point in time */
    instant!: Instant;
    /** Parameters with validity times over twelve hours. Will not exist for all time steps. */
    next_12_hours?: Next_12_hours | undefined;
    /** Parameters with validity times over one hour. Will not exist for all time steps. */
    next_1_hours?: Next_1_hours | undefined;
    /** Parameters with validity times over six hours. Will not exist for all time steps. */
    next_6_hours?: Next_6_hours | undefined;

    constructor(data?: IData) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.instant = new Instant();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.instant = _data["instant"] ? Instant.fromJS(_data["instant"]) : new Instant();
            this.next_12_hours = _data["next_12_hours"] ? Next_12_hours.fromJS(_data["next_12_hours"]) : <any>undefined;
            this.next_1_hours = _data["next_1_hours"] ? Next_1_hours.fromJS(_data["next_1_hours"]) : <any>undefined;
            this.next_6_hours = _data["next_6_hours"] ? Next_6_hours.fromJS(_data["next_6_hours"]) : <any>undefined;
        }
    }

    static fromJS(data: any): Data {
        data = typeof data === 'object' ? data : {};
        let result = new Data();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["instant"] = this.instant ? this.instant.toJSON() : <any>undefined;
        data["next_12_hours"] = this.next_12_hours ? this.next_12_hours.toJSON() : <any>undefined;
        data["next_1_hours"] = this.next_1_hours ? this.next_1_hours.toJSON() : <any>undefined;
        data["next_6_hours"] = this.next_6_hours ? this.next_6_hours.toJSON() : <any>undefined;
        return data;
    }
}

export interface IData {
    /** Parameters which applies to this exact point in time */
    instant: Instant;
    /** Parameters with validity times over twelve hours. Will not exist for all time steps. */
    next_12_hours?: Next_12_hours | undefined;
    /** Parameters with validity times over one hour. Will not exist for all time steps. */
    next_1_hours?: Next_1_hours | undefined;
    /** Parameters with validity times over six hours. Will not exist for all time steps. */
    next_6_hours?: Next_6_hours | undefined;
}

export enum METJSONForecastType {
    Feature = "Feature",
}

export enum PointGeometryType {
    Point = "Point",
}

export class Instant implements IInstant {
    details?: ForecastTimeInstant | undefined;

    constructor(data?: IInstant) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.details = _data["details"] ? ForecastTimeInstant.fromJS(_data["details"]) : <any>undefined;
        }
    }

    static fromJS(data: any): Instant {
        data = typeof data === 'object' ? data : {};
        let result = new Instant();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["details"] = this.details ? this.details.toJSON() : <any>undefined;
        return data;
    }
}

export interface IInstant {
    details?: ForecastTimeInstant | undefined;
}

export class Next_12_hours implements INext_12_hours {
    details!: ForecastTimePeriod;
    summary!: ForecastSummary;

    constructor(data?: INext_12_hours) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.details = new ForecastTimePeriod();
            this.summary = new ForecastSummary();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.details = _data["details"] ? ForecastTimePeriod.fromJS(_data["details"]) : new ForecastTimePeriod();
            this.summary = _data["summary"] ? ForecastSummary.fromJS(_data["summary"]) : new ForecastSummary();
        }
    }

    static fromJS(data: any): Next_12_hours {
        data = typeof data === 'object' ? data : {};
        let result = new Next_12_hours();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["details"] = this.details ? this.details.toJSON() : <any>undefined;
        data["summary"] = this.summary ? this.summary.toJSON() : <any>undefined;
        return data;
    }
}

export interface INext_12_hours {
    details: ForecastTimePeriod;
    summary: ForecastSummary;
}

export class Next_1_hours implements INext_1_hours {
    details!: ForecastTimePeriod;
    summary!: ForecastSummary;

    constructor(data?: INext_1_hours) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.details = new ForecastTimePeriod();
            this.summary = new ForecastSummary();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.details = _data["details"] ? ForecastTimePeriod.fromJS(_data["details"]) : new ForecastTimePeriod();
            this.summary = _data["summary"] ? ForecastSummary.fromJS(_data["summary"]) : new ForecastSummary();
        }
    }

    static fromJS(data: any): Next_1_hours {
        data = typeof data === 'object' ? data : {};
        let result = new Next_1_hours();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["details"] = this.details ? this.details.toJSON() : <any>undefined;
        data["summary"] = this.summary ? this.summary.toJSON() : <any>undefined;
        return data;
    }
}

export interface INext_1_hours {
    details: ForecastTimePeriod;
    summary: ForecastSummary;
}

export class Next_6_hours implements INext_6_hours {
    details!: ForecastTimePeriod;
    summary!: ForecastSummary;

    constructor(data?: INext_6_hours) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        if (!data) {
            this.details = new ForecastTimePeriod();
            this.summary = new ForecastSummary();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.details = _data["details"] ? ForecastTimePeriod.fromJS(_data["details"]) : new ForecastTimePeriod();
            this.summary = _data["summary"] ? ForecastSummary.fromJS(_data["summary"]) : new ForecastSummary();
        }
    }

    static fromJS(data: any): Next_6_hours {
        data = typeof data === 'object' ? data : {};
        let result = new Next_6_hours();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["details"] = this.details ? this.details.toJSON() : <any>undefined;
        data["summary"] = this.summary ? this.summary.toJSON() : <any>undefined;
        return data;
    }
}